sql: Move scrub indexCheckOperation to scrub_index.go
This commit only includes code movement and some additional
documentation to indexCheckOperation to provide more context for the
contents of the file.
